update query input interactor & connector
