+++ Sources/InstantSearchCore/QueryInput/QueryInputInteractor+Searcher.swift
@@ -1,22 +1,15 @@
 //
 //  QueryInputInteractor+Searcher.swift
-//  InstantSearchCore
+//  
 //
-//  Created by Vladislav Fitc on 28/05/2019.
-//  Copyright © 2019 Algolia. All rights reserved.
+//  Created by Vladislav Fitc on 13/08/2021.
 //
 
 import Foundation
 
-public enum SearchTriggeringMode {
-  case searchAsYouType
-  case searchOnSubmit
-}
-
 public extension QueryInputInteractor {
 
-  @available(*, deprecated, message: "Use QueryInputInteractor.TextualQuerySearcherConnection")
-  struct SearcherConnection<S: Searcher>: Connection {
+  struct SubscriberConnection<S: AnyObject & QuerySettable & Searchable>: Connection {
 
     public let interactor: QueryInputInteractor
     public let searcher: S
@@ -32,18 +25,16 @@ public extension QueryInputInteractor {
 
     public func connect() {
 
-      interactor.query = searcher.query
-
       switch searchTriggeringMode {
       case .searchAsYouType:
         interactor.onQueryChanged.subscribe(with: searcher) { searcher, query in
-          searcher.query = query
+          searcher.setQuery(query)
           searcher.search()
         }
 
       case .searchOnSubmit:
         interactor.onQuerySubmitted.subscribe(with: searcher) { searcher, query in
-          searcher.query = query
+          searcher.setQuery(query)
           searcher.search()
         }
       }
@@ -69,10 +60,9 @@ public extension QueryInputInteractor {
 
 public extension QueryInputInteractor {
 
-  @available(*, deprecated, message: "Use QueryInputInteractor.TextualQuerySearcherConnection")
-  @discardableResult func connectSearcher<S: Searcher>(_ searcher: S,
-                                                       searchTriggeringMode: SearchTriggeringMode = .searchAsYouType) -> SearcherConnection<S> {
-    let connection = SearcherConnection(interactor: self, searcher: searcher, searchTriggeringMode: searchTriggeringMode)
+  @discardableResult func connectSearcher<S: AnyObject & QuerySettable>(_ searcher: S,
+                                                                        searchTriggeringMode: SearchTriggeringMode = .searchAsYouType) -> SubscriberConnection<S> {
+    let connection = SubscriberConnection(interactor: self, searcher: searcher, searchTriggeringMode: searchTriggeringMode)
     connection.connect()
     return connection
   }
