+++ Sources/InstantSearchCore/QueryInput/SearchTriggeringMode.swift
@@ -1,0 +1,13 @@
+//
+//  SearchTriggeringMode.swift
+//  
+//
+//  Created by Vladislav Fitc on 11/09/2021.
+//
+
+import Foundation
+
+public enum SearchTriggeringMode {
+  case searchAsYouType
+  case searchOnSubmit
+}
